requirements.txt
aiogram
yt-dlp
requests
gunicorn
flask
python-dotenv
